Move instance Unifiable to Malgo.Unify from Malgo.TypeRep.UTerm
